Rohan | Changed the quantity type from Integer to Double

+++ api-1.10/src/main/java/org/openmrs/module/emrapi/encounter/mapper/OrderMapper1_10.java
@@ -98,7 +98,7 @@ public class OrderMapper1_10 implements OrderMapper {
             dosingInstructions.setFrequency(openMRSDrugOrder.getFrequency().getName());
         }
         if (openMRSDrugOrder.getQuantity() != null) {
-            dosingInstructions.setQuantity(openMRSDrugOrder.getQuantity().intValue());
+            dosingInstructions.setQuantity(openMRSDrugOrder.getQuantity());
         }
         dosingInstructions.setQuantityUnits(getConceptName(openMRSDrugOrder.getQuantityUnits()));
         dosingInstructions.setAdministrationInstructions(openMRSDrugOrder.getDosingInstructions());
